Fixed install paths on *NIX
The previous build was hard to configure for packagers. This one uses a CMake module (`GNUInstallDirs`) to standardize it so that it works like most programs out there. I already tested this both on my local machine and on the Open Build System (openSUSE).

The Mac part might need some fixing, I had to use the `INSTALL_DIR` variable for consistency but the logic should be the same as before.